BMI index for add details can calculated automotically (no input for that field is needed)
Note that BMVS table will need to be migrated.
Data type changed in BMVS migrations from int to decimal:
weight, height, BMI, temp.

BMI can be calculated automatically but will not show up immediately as it will be calculated after user enter create button

+++ database/migrations/2020_06_07_082107_create_patient__body_and_vital_signs_table.php
@@ -18,10 +18,10 @@ class CreatePatientBodyAndVitalSignsTable extends Migration
             $table->integer('patient_id');
             $table->date('dateTaken');
             $table->time('timeTaken');
-            $table->integer('weight');
-            $table->integer('height');
-            $table->integer('bmi');
-            $table->integer('temperature');
+            $table->decimal('weight');
+            $table->decimal('height');
+            $table->decimal('bmi');
+            $table->decimal('temperature');
             $table->integer('Supine_ReadingTime');
             $table->integer('Supine_BP');
             $table->integer('Supine_HR');
